+++ rust/dbn/src/publishers.rs
@@ -37,7 +37,7 @@ pub enum Venue {
     Xase = 11,
     /// NYSE Arca
     Arcx = 12,
-    /// NYSE Chicago, Inc.
+    /// NYSE Texas, Inc.
     Xchi = 13,
     /// Investors Exchange
     Iexg = 14,
@@ -53,21 +53,21 @@ pub enum Venue {
     Eprl = 19,
     /// NYSE American Options
     Amxo = 20,
-    /// BOX Options Exchange
+    /// BOX Options
     Xbox = 21,
-    /// Cboe Options Exchange
+    /// Cboe Options
     Xcbo = 22,
     /// MIAX Emerald
     Emld = 23,
-    /// Cboe EDGX Options Exchange
+    /// Cboe EDGX Options
     Edgo = 24,
-    /// ISE Gemini Exchange
+    /// Nasdaq GEMX
     Gmni = 25,
-    /// International Securities Exchange, LLC
+    /// Nasdaq ISE
     Xisx = 26,
-    /// ISE Mercury, LLC
+    /// Nasdaq MRX
     Mcry = 27,
-    /// Miami International Securities Exchange
+    /// MIAX Options
     Xmio = 28,
     /// NYSE Arca Options
     Arco = 29,
@@ -75,17 +75,17 @@ pub enum Venue {
     Opra = 30,
     /// MIAX Pearl
     Mprl = 31,
-    /// Nasdaq Options Market
+    /// Nasdaq Options
     Xndq = 32,
-    /// Nasdaq OMX BX Options
+    /// Nasdaq BX Options
     Xbxo = 33,
-    /// Cboe C2 Options Exchange
+    /// Cboe C2 Options
     C2Ox = 34,
-    /// Nasdaq OMX PHLX
+    /// Nasdaq PHLX
     Xphl = 35,
-    /// Cboe BZX Options Exchange
+    /// Cboe BZX Options
     Bato = 36,
-    /// MEMX LLC Options
+    /// MEMX Options
     Mxop = 37,
     /// ICE Futures Europe (Commodities)
     Ifeu = 38,
@@ -281,7 +281,7 @@ pub enum Dataset {
     XcisPillar = 10,
     /// NYSE American Integrated
     XasePillar = 11,
-    /// NYSE Chicago Integrated
+    /// NYSE Texas Integrated
     XchiPillar = 12,
     /// NYSE National BBO
     XcisBbo = 13,
@@ -485,7 +485,7 @@ pub enum Publisher {
     XcisPillarXcis = 10,
     /// NYSE American Integrated
     XasePillarXase = 11,
-    /// NYSE Chicago Integrated
+    /// NYSE Texas Integrated
     XchiPillarXchi = 12,
     /// NYSE National BBO
     XcisBboXcis = 13,
@@ -501,15 +501,15 @@ pub enum Publisher {
     XnasNlsFinc = 18,
     /// FINRA/NYSE TRF
     XnysTradesFiny = 19,
-    /// OPRA - NYSE American
+    /// OPRA - NYSE American Options
     OpraPillarAmxo = 20,
-    /// OPRA - Boston Options Exchange
+    /// OPRA - BOX Options
     OpraPillarXbox = 21,
-    /// OPRA - Cboe Options Exchange
+    /// OPRA - Cboe Options
     OpraPillarXcbo = 22,
     /// OPRA - MIAX Emerald
     OpraPillarEmld = 23,
-    /// OPRA - Cboe EDGX Options Exchange
+    /// OPRA - Cboe EDGX Options
     OpraPillarEdgo = 24,
     /// OPRA - Nasdaq GEMX
     OpraPillarGmni = 25,
@@ -517,29 +517,29 @@ pub enum Publisher {
     OpraPillarXisx = 26,
     /// OPRA - Nasdaq MRX
     OpraPillarMcry = 27,
-    /// OPRA - Miami International Securities
+    /// OPRA - MIAX Options
     OpraPillarXmio = 28,
-    /// OPRA - NYSE Arca
+    /// OPRA - NYSE Arca Options
     OpraPillarArco = 29,
     /// OPRA - Options Price Reporting Authority
     OpraPillarOpra = 30,
     /// OPRA - MIAX Pearl
     OpraPillarMprl = 31,
-    /// OPRA - Nasdaq Options Market
+    /// OPRA - Nasdaq Options
     OpraPillarXndq = 32,
     /// OPRA - Nasdaq BX Options
     OpraPillarXbxo = 33,
-    /// OPRA - Cboe C2 Options Exchange
+    /// OPRA - Cboe C2 Options
     OpraPillarC2Ox = 34,
     /// OPRA - Nasdaq PHLX
     OpraPillarXphl = 35,
     /// OPRA - Cboe BZX Options
     OpraPillarBato = 36,
-    /// OPRA - MEMX Options Exchange
+    /// OPRA - MEMX Options
     OpraPillarMxop = 37,
     /// IEX TOPS
     IexgTopsIexg = 38,
-    /// DBEQ Basic - NYSE Chicago
+    /// DBEQ Basic - NYSE Texas
     DbeqBasicXchi = 39,
     /// DBEQ Basic - NYSE National
     DbeqBasicXcis = 40,
@@ -557,7 +557,7 @@ pub enum Publisher {
     XnasQbboXnas = 46,
     /// Nasdaq Trades
     XnasNlsXnas = 47,
-    /// Databento US Equities Plus - NYSE Chicago
+    /// Databento US Equities Plus - NYSE Texas
     EqusPlusXchi = 48,
     /// Databento US Equities Plus - NYSE National
     EqusPlusXcis = 49,
@@ -585,7 +585,7 @@ pub enum Publisher {
     EqusPlusEqus = 60,
     /// OPRA - MIAX Sapphire
     OpraPillarSphr = 61,
-    /// Databento US Equities (All Feeds) - NYSE Chicago
+    /// Databento US Equities (All Feeds) - NYSE Texas
     EqusAllXchi = 62,
     /// Databento US Equities (All Feeds) - NYSE National
     EqusAllXcis = 63,
@@ -667,10 +667,14 @@ pub enum Publisher {
     XeurEobiXeur = 101,
     /// European Energy Exchange EOBI
     XeerEobiXeer = 102,
+    /// Eurex EOBI - Off-Market Trades
+    XeurEobiXoff = 103,
+    /// European Energy Exchange EOBI - Off-Market Trades
+    XeerEobiXoff = 104,
 }
 
 /// The number of Publisher variants.
-pub const PUBLISHER_COUNT: usize = 102;
+pub const PUBLISHER_COUNT: usize = 104;
 
 impl Publisher {
     /// Convert a Publisher to its `str` representation.
@@ -778,6 +782,8 @@ impl Publisher {
             Self::IfllImpactXoff => "IFLL.IMPACT.XOFF",
             Self::XeurEobiXeur => "XEUR.EOBI.XEUR",
             Self::XeerEobiXeer => "XEER.EOBI.XEER",
+            Self::XeurEobiXoff => "XEUR.EOBI.XOFF",
+            Self::XeerEobiXoff => "XEER.EOBI.XOFF",
         }
     }
 
@@ -886,6 +892,8 @@ impl Publisher {
             Self::IfllImpactXoff => Venue::Xoff,
             Self::XeurEobiXeur => Venue::Xeur,
             Self::XeerEobiXeer => Venue::Xeer,
+            Self::XeurEobiXoff => Venue::Xoff,
+            Self::XeerEobiXoff => Venue::Xoff,
         }
     }
 
@@ -994,6 +1002,8 @@ impl Publisher {
             Self::IfllImpactXoff => Dataset::IfllImpact,
             Self::XeurEobiXeur => Dataset::XeurEobi,
             Self::XeerEobiXeer => Dataset::XeerEobi,
+            Self::XeurEobiXoff => Dataset::XeurEobi,
+            Self::XeerEobiXoff => Dataset::XeerEobi,
         }
     }
 
@@ -1104,6 +1114,8 @@ impl Publisher {
             (Dataset::IfllImpact, Venue::Xoff) => Ok(Self::IfllImpactXoff),
             (Dataset::XeurEobi, Venue::Xeur) => Ok(Self::XeurEobiXeur),
             (Dataset::XeerEobi, Venue::Xeer) => Ok(Self::XeerEobiXeer),
+            (Dataset::XeurEobi, Venue::Xoff) => Ok(Self::XeurEobiXoff),
+            (Dataset::XeerEobi, Venue::Xoff) => Ok(Self::XeerEobiXoff),
             _ => Err(Error::conversion::<Self>(format!("({dataset}, {venue})"))),
         }
     }
@@ -1228,6 +1240,8 @@ impl std::str::FromStr for Publisher {
             "IFLL.IMPACT.XOFF" => Ok(Self::IfllImpactXoff),
             "XEUR.EOBI.XEUR" => Ok(Self::XeurEobiXeur),
             "XEER.EOBI.XEER" => Ok(Self::XeerEobiXeer),
+            "XEUR.EOBI.XOFF" => Ok(Self::XeurEobiXoff),
+            "XEER.EOBI.XOFF" => Ok(Self::XeerEobiXoff),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }