+++ rust/dbn/src/v3/impl_default.rs
@@ -1,11 +1,11 @@
 use std::os::raw::c_char;
 
 use crate::{
-    rtype, MatchAlgorithm, RecordHeader, SecurityUpdateAction, Side, UserDefinedInstrument,
-    UNDEF_PRICE, UNDEF_TIMESTAMP,
+    rtype, MatchAlgorithm, RecordHeader, SecurityUpdateAction, Side, StatUpdateAction,
+    UserDefinedInstrument, UNDEF_PRICE, UNDEF_TIMESTAMP,
 };
 
-use super::{InstrumentDefMsg, SYMBOL_CSTR_LEN};
+use super::{InstrumentDefMsg, StatMsg, SYMBOL_CSTR_LEN, UNDEF_STAT_QUANTITY};
 
 impl Default for InstrumentDefMsg {
     fn default() -> Self {
@@ -80,6 +80,25 @@ impl Default for InstrumentDefMsg {
             leg_raw_symbol: [0; SYMBOL_CSTR_LEN],
             leg_instrument_class: 0,
             leg_side: Side::None as c_char,
+            _reserved: Default::default(),
+        }
+    }
+}
+
+impl Default for StatMsg {
+    fn default() -> Self {
+        Self {
+            hd: RecordHeader::default::<Self>(rtype::STATISTICS),
+            ts_recv: UNDEF_TIMESTAMP,
+            ts_ref: UNDEF_TIMESTAMP,
+            price: UNDEF_PRICE,
+            quantity: UNDEF_STAT_QUANTITY,
+            sequence: 0,
+            ts_in_delta: 0,
+            stat_type: 0,
+            channel_id: 0,
+            update_action: StatUpdateAction::New as u8,
+            stat_flags: 0,
             _reserved: Default::default(),
         }
     }